fix: address codex + gemini findings on PR #46 language-shim experiment
Four valid findings, all addressed:

CODEX P1 (real bug, set_camera_transform partial-update destruction):
  Original: missing 'location' or 'rotation' fields defaulted to 0.0,
  silently snapping the omitted side to the world origin. Calls supplying
  only one side were destructive.
  Fix: when an omitted side is detected, run get_camera_transform first
  to read the current value, then forward the full set call. Two-round-trip
  cost on partial updates is the price of correctness; full updates (both
  sides supplied) skip the read entirely. Empty-args is now a no-op read.
  Response includes a 'preserved' map indicating which sides were omitted.

CODEX P2 + GEMINI medium (same finding, marker-search log window):
  get_camera_transform fetched only 200 LogPython lines but the LogCapture
  ring holds up to 1000. >200 log lines between exec and read could
  silently lose the marker. Fix: fetch full ring (1000 lines).

GEMINI medium (NaN/Infinity in _num):
  Original validation accepted any int/float, including NaN and Infinity
  -- which would generate malformed Python like 'unreal.Vector(nan, ...)'
  and either crash UE or set invalid state. Fix: reject non-finite values
  via math.isfinite(). Added 'import math'.

GEMINI medium (int64 accumulators in inspect_static_mesh):
  Per-LOD vertex/triangle counts are int32; the total across many LODs
  on a high-poly mesh could theoretically exceed INT32_MAX. Cheap
  defensive change to int64 accumulators.

DISMISSED: Gemini medium (use LoadObject<UObject> instead of
UEditorAssetLibrary::LoadAsset). Every existing inspect/compile/move/
rename/delete handler in this codebase uses UEditorAssetLibrary;
EditorScriptingUtilities is already a Build.cs dep. Switching just one
handler would create a precedent inconsistency for marginal benefit.
Per directive #4 in HANDOFF.md, my judgment with source-grounded
reasoning wins.

Tests: 143 still green.

+++ UnrealClaudeMCP/Source/UnrealClaudeMCP/Private/MCP/Handlers/Handler_InspectStaticMesh.cpp
@@ -85,8 +85,11 @@ public:
         const int32 NumLODs = Mesh->GetNumLODs();
         TArray<TSharedPtr<FJsonValue>> LodArray;
         LodArray.Reserve(NumLODs);
-        int32 TotalVertices = 0;
-        int32 TotalTriangles = 0;
+        // int64 accumulators -- per-LOD counts are int32, but the total
+        // across many LODs on a high-poly mesh could theoretically exceed
+        // INT32_MAX. Cheap defensive change. (Gemini medium on PR #46.)
+        int64 TotalVertices = 0;
+        int64 TotalTriangles = 0;
         for (int32 i = 0; i < NumLODs; ++i)
         {
             const int32 V = Mesh->GetNumVertices(i);
